distance cast down to the nearest int, making it ever so slightly easier to cap the ghost

+++ mp/src/game/shared/neo/neo_ghost_cap_point.cpp
@@ -224,7 +224,7 @@ void CNEOGhostCapturePoint::Think_CheckMyRadius(void)
 		}
 
 		const Vector dir = player->GetAbsOrigin() - GetAbsOrigin();
-		const float distance = dir.Length();
+		const int distance = dir.Length();
 
 		Assert(distance >= 0);
